Updates plugins-api version and refactors package names to be similar to what was done in the engine seed.
