+++ src/main/scala/com/codacy/duplication/scala/seed/traits/Timeoutable.scala
@@ -1,4 +1,4 @@
-package codacy.docker.api.utils
+package com.codacy.duplication.scala.seed.traits
 
 import java.util.{Timer, TimerTask}
