+++ src/main/scala/com/codacy/duplication/scala/seed/utils/FileHelper.scala
@@ -1,4 +1,4 @@
-package codacy.docker.api.utils
+package com.codacy.duplication.scala.seed.utils
 
 import java.io.File
 import java.nio.charset.StandardCharsets
