+++ src/test/scala/com/codacy/duplication/scala/seed/traits/TimeoutableSpecs.scala
@@ -1,4 +1,4 @@
-package codacy.docker.api.utils
+package com.codacy.duplication.scala.seed.traits
 
 import org.specs2.mutable.Specification
